+++ src/screens/learning/components/FlashCard.js
