+++ src/screens/search/components/DictionaryEntriesContainer.js
