+++ src/screens/search/components/DictionaryEntry.js
