+++ src/screens/settings/components/NameBar.js
