+++ src/screens/settings/components/NameBarContainer.js
